@@ -1,12 +1,13 @@
-import { Notice, Plugin } from 'obsidian';
+import { Plugin } from 'obsidian';
 import { KhojSetting, KhojSettingTab, DEFAULT_SETTINGS } from 'src/settings'
 import { KhojSearchModal } from 'src/search_modal'
 import { KhojChatModal } from 'src/chat_modal'
-import { configureKhojBackend } from './utils';
+import { updateContentIndex, canConnectToBackend } from './utils';
 
 
 export default class Khoj extends Plugin {
     settings: KhojSetting;
+    indexingTimer: NodeJS.Timeout;
 
     async onload() {
         await this.loadSettings();
@@ -15,61 +16,57 @@ export default class Khoj extends Plugin {
         this.addCommand({
             id: 'search',
             name: 'Search',
-            checkCallback: (checking) => {
-                if (!checking && this.settings.connectedToBackend)
-                    new KhojSearchModal(this.app, this.settings).open();
-                return this.settings.connectedToBackend;
-            }
+            callback: () => { new KhojSearchModal(this.app, this.settings).open(); }
         });
 
         // Add similar notes command. It can only be triggered from the editor
         this.addCommand({
             id: 'similar',
             name: 'Find similar notes',
-            editorCheckCallback: (checking) => {
-                if (!checking && this.settings.connectedToBackend)
-                    new KhojSearchModal(this.app, this.settings, true).open();
-                return this.settings.connectedToBackend;
-            }
+            editorCallback: () => { new KhojSearchModal(this.app, this.settings, true).open(); }
         });
 
         // Add chat command. It can be triggered from anywhere
         this.addCommand({
             id: 'chat',
             name: 'Chat',
-            checkCallback: (checking) => {
-                if (!checking && this.settings.connectedToBackend && (!!this.settings.openaiApiKey || this.settings.enableOfflineChat))
-                    new KhojChatModal(this.app, this.settings).open();
-                return !!this.settings.openaiApiKey || this.settings.enableOfflineChat;
-            }
+            callback: () => { new KhojChatModal(this.app, this.settings).open(); }
         });
 
         // Create an icon in the left ribbon.
-        this.addRibbonIcon('search', 'Khoj', (_: MouseEvent) => {
-            // Called when the user clicks the icon.
-            this.settings.connectedToBackend
-                ? new KhojSearchModal(this.app, this.settings).open()
-                : new Notice(`❗️Ensure Khoj backend is running and Khoj URL is pointing to it in the plugin settings`);
+        this.addRibbonIcon('message-circle', 'Khoj', (_: MouseEvent) => {
+            new KhojChatModal(this.app, this.settings).open()
         });
 
         // Add a settings tab so the user can configure khoj
         this.addSettingTab(new KhojSettingTab(this.app, this));
+
+        // Add scheduled job to update index every 60 minutes
+        this.indexingTimer = setInterval(async () => {
+            if (this.settings.autoConfigure) {
+                this.settings.lastSync = await updateContentIndex(this.app.vault, this.settings, this.settings.lastSync);
+            }
+        }, 60 * 60 * 1000);
     }
 
     async loadSettings() {
         // Load khoj obsidian plugin settings
         this.settings = Object.assign({}, DEFAULT_SETTINGS, await this.loadData());
 
-        if (this.settings.autoConfigure) {
-            // Load, configure khoj server settings
-            await configureKhojBackend(this.app.vault, this.settings);
-        }
+        // Check if can connect to khoj server
+        ({ connectedToBackend: this.settings.connectedToBackend } =
+            await canConnectToBackend(this.settings.khojUrl, this.settings.khojApiKey, true));
     }
 
     async saveSettings() {
-        if (this.settings.autoConfigure) {
-            await configureKhojBackend(this.app.vault, this.settings, false);
-        }
         this.saveData(this.settings);
+    }
+
+    async onunload() {
+        // Remove scheduled job to update index at regular cadence
+        if (this.indexingTimer)
+            clearInterval(this.indexingTimer);
+
+        this.unload();
     }
 }
