@@ -1,13 +1,8 @@
-# Standard Packages
-import os
 import logging
-from typing import Any
+import os
 from threading import Thread
+from typing import Dict
 
-# External Packages
-from langchain.chat_models import ChatOpenAI
-from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
-from langchain.callbacks.base import BaseCallbackManager
 import openai
 from tenacity import (
     before_sleep_log,
@@ -18,50 +13,66 @@ from tenacity import (
     wait_random_exponential,
 )
 
-# Internal Packages
 from khoj.processor.conversation.utils import ThreadedGenerator
-
 
 logger = logging.getLogger(__name__)
 
-
-class StreamingChatCallbackHandler(StreamingStdOutCallbackHandler):
-    def __init__(self, gen: ThreadedGenerator):
-        super().__init__()
-        self.gen = gen
-
-    def on_llm_new_token(self, token: str, **kwargs) -> Any:
-        self.gen.send(token)
+openai_clients: Dict[str, openai.OpenAI] = {}
 
 
 @retry(
     retry=(
-        retry_if_exception_type(openai.error.Timeout)
-        | retry_if_exception_type(openai.error.APIError)
-        | retry_if_exception_type(openai.error.APIConnectionError)
-        | retry_if_exception_type(openai.error.RateLimitError)
-        | retry_if_exception_type(openai.error.ServiceUnavailableError)
+        retry_if_exception_type(openai._exceptions.APITimeoutError)
+        | retry_if_exception_type(openai._exceptions.APIError)
+        | retry_if_exception_type(openai._exceptions.APIConnectionError)
+        | retry_if_exception_type(openai._exceptions.RateLimitError)
+        | retry_if_exception_type(openai._exceptions.APIStatusError)
     ),
     wait=wait_random_exponential(min=1, max=10),
     stop=stop_after_attempt(3),
     before_sleep=before_sleep_log(logger, logging.DEBUG),
     reraise=True,
 )
-def completion_with_backoff(**kwargs):
-    messages = kwargs.pop("messages")
-    if not "openai_api_key" in kwargs:
-        kwargs["openai_api_key"] = os.getenv("OPENAI_API_KEY")
-    llm = ChatOpenAI(**kwargs, request_timeout=20, max_retries=1)
-    return llm(messages=messages)
+def completion_with_backoff(
+    messages, model, temperature=0, openai_api_key=None, api_base_url=None, model_kwargs=None
+) -> str:
+    client_key = f"{openai_api_key}--{api_base_url}"
+    client: openai.OpenAI = openai_clients.get(client_key)
+    if not client:
+        client = openai.OpenAI(
+            api_key=openai_api_key or os.getenv("OPENAI_API_KEY"),
+            base_url=api_base_url,
+        )
+        openai_clients[client_key] = client
+
+    formatted_messages = [{"role": message.role, "content": message.content} for message in messages]
+
+    chat = client.chat.completions.create(
+        stream=True,
+        messages=formatted_messages,  # type: ignore
+        model=model,  # type: ignore
+        temperature=temperature,
+        timeout=20,
+        **(model_kwargs or dict()),
+    )
+    aggregated_response = ""
+    for chunk in chat:
+        delta_chunk = chunk.choices[0].delta  # type: ignore
+        if isinstance(delta_chunk, str):
+            aggregated_response += delta_chunk
+        elif delta_chunk.content:
+            aggregated_response += delta_chunk.content
+
+    return aggregated_response
 
 
 @retry(
     retry=(
-        retry_if_exception_type(openai.error.Timeout)
-        | retry_if_exception_type(openai.error.APIError)
-        | retry_if_exception_type(openai.error.APIConnectionError)
-        | retry_if_exception_type(openai.error.RateLimitError)
-        | retry_if_exception_type(openai.error.ServiceUnavailableError)
+        retry_if_exception_type(openai._exceptions.APITimeoutError)
+        | retry_if_exception_type(openai._exceptions.APIError)
+        | retry_if_exception_type(openai._exceptions.APIConnectionError)
+        | retry_if_exception_type(openai._exceptions.RateLimitError)
+        | retry_if_exception_type(openai._exceptions.APIStatusError)
     ),
     wait=wait_exponential(multiplier=1, min=4, max=10),
     stop=stop_after_attempt(3),
@@ -69,33 +80,51 @@ def completion_with_backoff(**kwargs):
     reraise=True,
 )
 def chat_completion_with_backoff(
-    messages, compiled_references, model_name, temperature, openai_api_key=None, completion_func=None
+    messages,
+    compiled_references,
+    online_results,
+    model_name,
+    temperature,
+    openai_api_key=None,
+    api_base_url=None,
+    completion_func=None,
+    model_kwargs=None,
 ):
-    g = ThreadedGenerator(compiled_references, completion_func=completion_func)
-    t = Thread(target=llm_thread, args=(g, messages, model_name, temperature, openai_api_key))
+    g = ThreadedGenerator(compiled_references, online_results, completion_func=completion_func)
+    t = Thread(
+        target=llm_thread, args=(g, messages, model_name, temperature, openai_api_key, api_base_url, model_kwargs)
+    )
     t.start()
     return g
 
 
-def llm_thread(g, messages, model_name, temperature, openai_api_key=None):
-    callback_handler = StreamingChatCallbackHandler(g)
-    chat = ChatOpenAI(
-        streaming=True,
-        verbose=True,
-        callback_manager=BaseCallbackManager([callback_handler]),
-        model_name=model_name,  # type: ignore
+def llm_thread(g, messages, model_name, temperature, openai_api_key=None, api_base_url=None, model_kwargs=None):
+    client_key = f"{openai_api_key}--{api_base_url}"
+    if client_key not in openai_clients:
+        client: openai.OpenAI = openai.OpenAI(
+            api_key=openai_api_key or os.getenv("OPENAI_API_KEY"),
+            base_url=api_base_url,
+        )
+        openai_clients[client_key] = client
+    else:
+        client: openai.OpenAI = openai_clients[client_key]
+
+    formatted_messages = [{"role": message.role, "content": message.content} for message in messages]
+
+    chat = client.chat.completions.create(
+        stream=True,
+        messages=formatted_messages,
+        model=model_name,  # type: ignore
         temperature=temperature,
-        openai_api_key=openai_api_key or os.getenv("OPENAI_API_KEY"),
-        request_timeout=20,
-        max_retries=1,
-        client=None,
+        timeout=20,
+        **(model_kwargs or dict()),
     )
 
-    chat(messages=messages)
+    for chunk in chat:
+        delta_chunk = chunk.choices[0].delta
+        if isinstance(delta_chunk, str):
+            g.send(delta_chunk)
+        elif delta_chunk.content:
+            g.send(delta_chunk.content)
 
     g.close()
-
-
-def extract_summaries(metadata):
-    """Extract summaries from metadata"""
-    return "".join([f'\n{session["summary"]}' for session in metadata])
