@@ -1,23 +1,31 @@
-# Standard Packages
 import argparse
+import logging
+import os
 import pathlib
 from importlib.metadata import version
 
-# Internal Packages
-from khoj.utils.helpers import resolve_absolute_path
-from khoj.utils.yaml import parse_config_from_file
-from khoj.migrations.migrate_version import migrate_config_to_version
-from khoj.migrations.migrate_processor_config_openai import migrate_processor_conversation_schema
+logger = logging.getLogger(__name__)
+
+from khoj.migrations.migrate_offline_chat_default_model import (
+    migrate_offline_chat_default_model,
+)
+from khoj.migrations.migrate_offline_chat_schema import migrate_offline_chat_schema
 from khoj.migrations.migrate_offline_model import migrate_offline_model
+from khoj.migrations.migrate_processor_config_openai import (
+    migrate_processor_conversation_schema,
+)
+from khoj.migrations.migrate_server_pg import migrate_server_pg
+from khoj.migrations.migrate_version import migrate_config_to_version
+from khoj.utils.helpers import in_debug_mode, resolve_absolute_path
+from khoj.utils.yaml import parse_config_from_file
 
 
 def cli(args=None):
     # Setup Argument Parser for the Commandline Interface
     parser = argparse.ArgumentParser(description="Start Khoj; An AI personal assistant for your Digital Brain")
     parser.add_argument(
-        "--config-file", "-c", default="~/.khoj/khoj.yml", type=pathlib.Path, help="YAML file to configure Khoj"
+        "--config-file", default="~/.khoj/khoj.yml", type=pathlib.Path, help="YAML file to configure Khoj"
     )
-    parser.add_argument("--gui", action="store_true", default=False, help="Show native desktop GUI. Default: false")
     parser.add_argument(
         "--regenerate",
         action="store_true",
@@ -33,9 +41,23 @@ def cli(args=None):
         help="Path to UNIX socket for server. Use to run server behind reverse proxy. Default: /tmp/uvicorn.sock",
     )
     parser.add_argument("--version", "-V", action="store_true", help="Print the installed Khoj version and exit")
-    parser.add_argument("--demo", action="store_true", default=False, help="Run Khoj in demo mode")
+    parser.add_argument(
+        "--disable-chat-on-gpu", action="store_true", default=False, help="Disable using GPU for the offline chat model"
+    )
+    parser.add_argument(
+        "--anonymous-mode",
+        action="store_true",
+        default=False,
+        help="Run Khoj in anonymous mode. This does not require any login for connecting users.",
+    )
 
-    args = parser.parse_args(args)
+    args, remaining_args = parser.parse_known_args(args)
+
+    if len(remaining_args) > 0:
+        logger.info(f"⚠️  Ignoring unknown commandline args: {remaining_args}")
+
+    # Set default values for arguments
+    args.chat_on_gpu = not args.disable_chat_on_gpu
 
     args.version_no = version("khoj-assistant")
     if args.version:
@@ -51,12 +73,21 @@ def cli(args=None):
     else:
         args = run_migrations(args)
         args.config = parse_config_from_file(args.config_file)
+        if in_debug_mode():
+            args.config.app.should_log_telemetry = False
 
     return args
 
 
 def run_migrations(args):
-    migrations = [migrate_config_to_version, migrate_processor_conversation_schema, migrate_offline_model]
+    migrations = [
+        migrate_config_to_version,
+        migrate_processor_conversation_schema,
+        migrate_offline_model,
+        migrate_offline_chat_schema,
+        migrate_offline_chat_default_model,
+        migrate_server_pg,
+    ]
     for migration in migrations:
         args = migration(args)
     return args
