@@ -1,25 +1,22 @@
 # System Packages
 from __future__ import annotations  # to avoid quoting type hints
 
-from enum import Enum
 import logging
-
 from dataclasses import dataclass
-from pathlib import Path
-from typing import TYPE_CHECKING, Dict, List, Optional, Union, Any
-from khoj.processor.conversation.gpt4all.utils import download_model
+from enum import Enum
+from typing import TYPE_CHECKING, Any, List, Optional, Union
 
-# External Packages
 import torch
+
+from khoj.processor.conversation.offline.utils import download_model
 
 logger = logging.getLogger(__name__)
 
-# Internal Packages
+
 if TYPE_CHECKING:
     from sentence_transformers import CrossEncoder
-    from khoj.search_filter.base_filter import BaseFilter
+
     from khoj.utils.models import BaseEncoder
-    from khoj.utils.rawconfig import ConversationProcessorConfig, Entry, OpenAIProcessorConfig
 
 
 class SearchType(str, Enum):
@@ -39,9 +36,7 @@ class ProcessorType(str, Enum):
 
 @dataclass
 class TextContent:
-    entries: List[Entry]
-    corpus_embeddings: torch.Tensor
-    filters: List[BaseFilter]
+    enabled: bool
 
 
 @dataclass
@@ -64,53 +59,22 @@ class ImageSearchModel:
 
 
 @dataclass
-class ContentIndex:
-    org: Optional[TextContent] = None
-    markdown: Optional[TextContent] = None
-    pdf: Optional[TextContent] = None
-    github: Optional[TextContent] = None
-    notion: Optional[TextContent] = None
-    image: Optional[ImageContent] = None
-    plaintext: Optional[TextContent] = None
-    plugins: Optional[Dict[str, TextContent]] = None
-
-
-@dataclass
 class SearchModels:
     text_search: Optional[TextSearchModel] = None
-    image_search: Optional[ImageSearchModel] = None
-    plugin_search: Optional[Dict[str, TextSearchModel]] = None
 
 
 @dataclass
-class GPT4AllProcessorConfig:
-    chat_model: Optional[str] = "llama-2-7b-chat.ggmlv3.q4_K_S.bin"
+class OfflineChatProcessorConfig:
     loaded_model: Union[Any, None] = None
 
 
-class ConversationProcessorConfigModel:
-    def __init__(
-        self,
-        conversation_config: ConversationProcessorConfig,
-    ):
-        self.openai_model = conversation_config.openai
-        self.gpt4all_model = GPT4AllProcessorConfig()
-        self.enable_offline_chat = conversation_config.enable_offline_chat
-        self.conversation_logfile = Path(conversation_config.conversation_logfile)
-        self.chat_session: List[str] = []
-        self.meta_log: dict = {}
-
-        if self.enable_offline_chat:
-            try:
-                self.gpt4all_model.loaded_model = download_model(self.gpt4all_model.chat_model)
-            except ValueError as e:
-                self.gpt4all_model.loaded_model = None
-                logger.error(f"Error while loading offline chat model: {e}", exc_info=True)
-                self.enable_offline_chat = False
-        else:
-            self.gpt4all_model.loaded_model = None
-
-
-@dataclass
-class ProcessorConfigModel:
-    conversation: Union[ConversationProcessorConfigModel, None] = None
+class OfflineChatProcessorModel:
+    def __init__(self, chat_model: str = "NousResearch/Hermes-2-Pro-Mistral-7B-GGUF", max_tokens: int = None):
+        self.chat_model = chat_model
+        self.loaded_model = None
+        try:
+            self.loaded_model = download_model(self.chat_model, max_tokens=max_tokens)
+        except ValueError as e:
+            self.loaded_model = None
+            logger.error(f"Error while loading offline chat model: {e}", exc_info=True)
+            raise e
