@@ -1,19 +1,18 @@
 # System Packages
 import json
+import uuid
 from pathlib import Path
-from typing import List, Dict, Optional, Union, Any
+from typing import Dict, List, Optional
 
-# External Packages
-from pydantic import BaseModel, validator
+from pydantic import BaseModel
 
-# Internal Packages
-from khoj.utils.helpers import to_snake_case_from_dash, is_none_or_empty
+from khoj.utils.helpers import to_snake_case_from_dash
 
 
 class ConfigBase(BaseModel):
     class Config:
         alias_generator = to_snake_case_from_dash
-        allow_population_by_field_name = True
+        populate_by_name = True
 
     def __getitem__(self, item):
         return getattr(self, item)
@@ -22,23 +21,21 @@ class ConfigBase(BaseModel):
         return setattr(self, key, value)
 
 
+class LocationData(BaseModel):
+    city: Optional[str]
+    region: Optional[str]
+    country: Optional[str]
+
+
 class TextConfigBase(ConfigBase):
     compressed_jsonl: Path
     embeddings_file: Path
 
 
-class TextContentConfig(TextConfigBase):
-    input_files: Optional[List[Path]]
-    input_filter: Optional[List[str]]
+class TextContentConfig(ConfigBase):
+    input_files: Optional[List[Path]] = None
+    input_filter: Optional[List[str]] = None
     index_heading_entries: Optional[bool] = False
-
-    @validator("input_filter")
-    def input_filter_or_files_required(cls, input_filter, values, **kwargs):
-        if is_none_or_empty(input_filter) and ("input_files" not in values or values["input_files"] is None):
-            raise ValueError(
-                "Either input_filter or input_files required in all content-type.<text_search> section of Khoj config file"
-            )
-        return input_filter
 
 
 class GithubRepoConfig(ConfigBase):
@@ -47,61 +44,35 @@ class GithubRepoConfig(ConfigBase):
     branch: Optional[str] = "master"
 
 
-class GithubContentConfig(TextConfigBase):
+class GithubContentConfig(ConfigBase):
     pat_token: str
     repos: List[GithubRepoConfig]
 
 
-class NotionContentConfig(TextConfigBase):
+class NotionContentConfig(ConfigBase):
     token: str
 
 
-class ImageContentConfig(ConfigBase):
-    input_directories: Optional[List[Path]]
-    input_filter: Optional[List[str]]
-    embeddings_file: Path
-    use_xmp_metadata: bool
-    batch_size: int
-
-    @validator("input_filter")
-    def input_filter_or_directories_required(cls, input_filter, values, **kwargs):
-        if is_none_or_empty(input_filter) and (
-            "input_directories" not in values or values["input_directories"] is None
-        ):
-            raise ValueError(
-                "Either input_filter or input_directories required in all content-type.image section of Khoj config file"
-            )
-        return input_filter
-
-
 class ContentConfig(ConfigBase):
-    org: Optional[TextContentConfig]
-    image: Optional[ImageContentConfig]
-    markdown: Optional[TextContentConfig]
-    pdf: Optional[TextContentConfig]
-    plaintext: Optional[TextContentConfig]
-    github: Optional[GithubContentConfig]
-    plugins: Optional[Dict[str, TextContentConfig]]
-    notion: Optional[NotionContentConfig]
-
-
-class TextSearchConfig(ConfigBase):
-    encoder: str
-    cross_encoder: str
-    encoder_type: Optional[str]
-    model_directory: Optional[Path]
+    org: Optional[TextContentConfig] = None
+    markdown: Optional[TextContentConfig] = None
+    pdf: Optional[TextContentConfig] = None
+    plaintext: Optional[TextContentConfig] = None
+    github: Optional[GithubContentConfig] = None
+    notion: Optional[NotionContentConfig] = None
 
 
 class ImageSearchConfig(ConfigBase):
     encoder: str
-    encoder_type: Optional[str]
-    model_directory: Optional[Path]
+    encoder_type: Optional[str] = None
+    model_directory: Optional[Path] = None
+
+    class Config:
+        protected_namespaces = ()
 
 
 class SearchConfig(ConfigBase):
-    asymmetric: Optional[TextSearchConfig]
-    symmetric: Optional[TextSearchConfig]
-    image: Optional[ImageSearchConfig]
+    image: Optional[ImageSearchConfig] = None
 
 
 class OpenAIProcessorConfig(ConfigBase):
@@ -109,32 +80,39 @@ class OpenAIProcessorConfig(ConfigBase):
     chat_model: Optional[str] = "gpt-3.5-turbo"
 
 
+class OfflineChatProcessorConfig(ConfigBase):
+    chat_model: Optional[str] = "NousResearch/Hermes-2-Pro-Mistral-7B-GGUF"
+
+
 class ConversationProcessorConfig(ConfigBase):
-    conversation_logfile: Path
-    openai: Optional[OpenAIProcessorConfig]
-    enable_offline_chat: Optional[bool] = False
+    openai: Optional[OpenAIProcessorConfig] = None
+    offline_chat: Optional[OfflineChatProcessorConfig] = None
+    max_prompt_size: Optional[int] = None
+    tokenizer: Optional[str] = None
 
 
 class ProcessorConfig(ConfigBase):
-    conversation: Optional[ConversationProcessorConfig]
+    conversation: Optional[ConversationProcessorConfig] = None
 
 
 class AppConfig(ConfigBase):
-    should_log_telemetry: bool
+    should_log_telemetry: bool = True
 
 
 class FullConfig(ConfigBase):
     content_type: Optional[ContentConfig] = None
     search_type: Optional[SearchConfig] = None
     processor: Optional[ProcessorConfig] = None
-    app: Optional[AppConfig] = AppConfig(should_log_telemetry=True)
+    app: Optional[AppConfig] = AppConfig()
     version: Optional[str] = None
 
 
 class SearchResponse(ConfigBase):
     entry: str
-    score: str
-    additional: Optional[dict]
+    score: float
+    cross_score: Optional[float] = None
+    additional: Optional[dict] = None
+    corpus_id: str
 
 
 class Entry:
@@ -142,14 +120,21 @@ class Entry:
     compiled: str
     heading: Optional[str]
     file: Optional[str]
+    corpus_id: str
 
     def __init__(
-        self, raw: str = None, compiled: str = None, heading: Optional[str] = None, file: Optional[str] = None
+        self,
+        raw: str = None,
+        compiled: str = None,
+        heading: Optional[str] = None,
+        file: Optional[str] = None,
+        corpus_id: uuid.UUID = None,
     ):
         self.raw = raw
         self.compiled = compiled
         self.heading = heading
         self.file = file
+        self.corpus_id = str(corpus_id)
 
     def to_json(self) -> str:
         return json.dumps(self.__dict__, ensure_ascii=False)
@@ -164,4 +149,5 @@ class Entry:
             compiled=dictionary["compiled"],
             file=dictionary.get("file", None),
             heading=dictionary.get("heading", None),
+            corpus_id=dictionary.get("corpus_id", None),
         )
