@@ -1,41 +1,44 @@
-# Standard Packages
+import os
 import threading
-from typing import List, Dict
-from packaging import version
-
-# External Packages
-import torch
+from collections import defaultdict
 from pathlib import Path
+from typing import Any, Dict, List
 
-# Internal Packages
+from apscheduler.schedulers.background import BackgroundScheduler
+from openai import OpenAI
+from whisper import Whisper
+
+from khoj.processor.embeddings import CrossEncoderModel, EmbeddingsModel
 from khoj.utils import config as utils_config
-from khoj.utils.config import ContentIndex, SearchModels, ProcessorConfigModel
-from khoj.utils.helpers import LRU
+from khoj.utils.config import OfflineChatProcessorModel, SearchModels
+from khoj.utils.helpers import LRU, get_device
 from khoj.utils.rawconfig import FullConfig
 
 # Application Global State
 config = FullConfig()
 search_models = SearchModels()
-content_index = ContentIndex()
-processor_config = ProcessorConfigModel()
+embeddings_model: Dict[str, EmbeddingsModel] = None
+cross_encoder_model: Dict[str, CrossEncoderModel] = None
+openai_client: OpenAI = None
+offline_chat_processor_config: OfflineChatProcessorModel = None
+whisper_model: Whisper = None
 config_file: Path = None
 verbose: int = 0
 host: str = None
 port: int = None
 cli_args: List[str] = None
-query_cache = LRU()
-config_lock = threading.Lock()
+query_cache: Dict[str, LRU] = defaultdict(LRU)
 chat_lock = threading.Lock()
 SearchType = utils_config.SearchType
+scheduler: BackgroundScheduler = None
 telemetry: List[Dict[str, str]] = []
-previous_query: str = None
-demo: bool = False
-
-if torch.cuda.is_available():
-    # Use CUDA GPU
-    device = torch.device("cuda:0")
-elif version.parse(torch.__version__) >= version.parse("1.13.0.dev") and torch.backends.mps.is_available():
-    # Use Apple M1 Metal Acceleration
-    device = torch.device("mps")
-else:
-    device = torch.device("cpu")
+khoj_version: str = None
+device = get_device()
+chat_on_gpu: bool = True
+anonymous_mode: bool = False
+pretrained_tokenizers: Dict[str, Any] = dict()
+billing_enabled: bool = (
+    os.getenv("STRIPE_API_KEY") is not None
+    and os.getenv("STRIPE_SIGNING_SECRET") is not None
+    and os.getenv("KHOJ_CLOUD_SUBSCRIPTION_URL") is not None
+)
