@@ -1,17 +1,13 @@
-# Standard Packages
 import argparse
 import logging
 import os
+import sqlite3
 from typing import Dict, List
 
-# External Packages
-from fastapi import FastAPI
-from fastapi import HTTPException
-from posthog import Posthog
-from dotenv import load_dotenv
-import sqlite3
 import uvicorn
-
+from dotenv import load_dotenv
+from fastapi import FastAPI, HTTPException
+from posthog import Posthog
 
 # Initialize Global App Variables
 app = FastAPI()
@@ -47,7 +43,7 @@ def v1_telemetry(telemetry_data: List[Dict[str, str]]):
 
         # Create a table if it doesn't exist
         cur.execute(
-            """CREATE TABLE IF NOT EXISTS usage (id INTEGER PRIMARY KEY, time TIMESTAMP, type TEXT, server_id TEXT, os TEXT, api TEXT, client TEXT)"""
+            """CREATE TABLE IF NOT EXISTS usage (id INTEGER PRIMARY KEY, time TIMESTAMP, type TEXT, server_id TEXT, os TEXT, api TEXT, client TEXT, server_version TEXT)"""
         )
 
         # Log telemetry data
