@@ -1,53 +1,9 @@
-# Standard Packages
 import re
 from datetime import datetime
-from math import inf
 
-# External Packages
 import pytest
 
-# Internal Packages
 from khoj.search_filter.date_filter import DateFilter
-from khoj.utils.rawconfig import Entry
-
-
-@pytest.mark.filterwarnings("ignore:The localize method is no longer necessary.")
-def test_date_filter():
-    entries = [
-        Entry(compiled="", raw="Entry with no date"),
-        Entry(compiled="", raw="April Fools entry: 1984-04-01"),
-        Entry(compiled="", raw="Entry with date:1984-04-02"),
-    ]
-
-    q_with_no_date_filter = "head tail"
-    ret_query, entry_indices = DateFilter().apply(q_with_no_date_filter, entries)
-    assert ret_query == "head tail"
-    assert entry_indices == {0, 1, 2}
-
-    q_with_dtrange_non_overlapping_at_boundary = 'head dt>"1984-04-01" dt<"1984-04-02" tail'
-    ret_query, entry_indices = DateFilter().apply(q_with_dtrange_non_overlapping_at_boundary, entries)
-    assert ret_query == "head tail"
-    assert entry_indices == set()
-
-    query_with_overlapping_dtrange = 'head dt>"1984-04-01" dt<"1984-04-03" tail'
-    ret_query, entry_indices = DateFilter().apply(query_with_overlapping_dtrange, entries)
-    assert ret_query == "head tail"
-    assert entry_indices == {2}
-
-    query_with_overlapping_dtrange = 'head dt>="1984-04-01" dt<"1984-04-02" tail'
-    ret_query, entry_indices = DateFilter().apply(query_with_overlapping_dtrange, entries)
-    assert ret_query == "head tail"
-    assert entry_indices == {1}
-
-    query_with_overlapping_dtrange = 'head dt>"1984-04-01" dt<="1984-04-02" tail'
-    ret_query, entry_indices = DateFilter().apply(query_with_overlapping_dtrange, entries)
-    assert ret_query == "head tail"
-    assert entry_indices == {2}
-
-    query_with_overlapping_dtrange = 'head dt>="1984-04-01" dt<="1984-04-02" tail'
-    ret_query, entry_indices = DateFilter().apply(query_with_overlapping_dtrange, entries)
-    assert ret_query == "head tail"
-    assert entry_indices == {1, 2}
 
 
 @pytest.mark.filterwarnings("ignore:The localize method is no longer necessary.")
@@ -56,8 +12,8 @@ def test_extract_date_range():
         datetime(1984, 1, 5, 0, 0, 0).timestamp(),
         datetime(1984, 1, 7, 0, 0, 0).timestamp(),
     ]
-    assert DateFilter().extract_date_range('head dt<="1984-01-01"') == [0, datetime(1984, 1, 2, 0, 0, 0).timestamp()]
-    assert DateFilter().extract_date_range('head dt>="1984-01-01"') == [datetime(1984, 1, 1, 0, 0, 0).timestamp(), inf]
+    assert DateFilter().extract_date_range('head dt<="1984-01-01"') == [None, datetime(1984, 1, 2, 0, 0, 0).timestamp()]
+    assert DateFilter().extract_date_range('head dt>="1984-01-01"') == [datetime(1984, 1, 1, 0, 0, 0).timestamp(), None]
     assert DateFilter().extract_date_range('head dt:"1984-01-01"') == [
         datetime(1984, 1, 1, 0, 0, 0).timestamp(),
         datetime(1984, 1, 2, 0, 0, 0).timestamp(),
@@ -68,13 +24,13 @@ def test_extract_date_range():
     ]
 
     # Unparseable date filter specified in query
-    assert DateFilter().extract_date_range('head dt:"Summer of 69" tail') == None
+    assert DateFilter().extract_date_range('head dt:"Summer of 69" tail') == []
 
     # No date filter specified in query
-    assert DateFilter().extract_date_range("head tail") == None
+    assert DateFilter().extract_date_range("head tail") == []
 
     # Non intersecting date ranges
-    assert DateFilter().extract_date_range('head dt>"1984-01-01" dt<"1984-01-01" tail') == None
+    assert DateFilter().extract_date_range('head dt>"1984-01-01" dt<"1984-01-01" tail') == []
 
 
 @pytest.mark.filterwarnings("ignore:The localize method is no longer necessary.")
@@ -158,3 +114,103 @@ def test_date_filter_regex():
 
     dtrange_match = re.findall(DateFilter().date_regex, "head tail")
     assert dtrange_match == []
+
+
+def test_get_date_filter_terms():
+    dtrange_match = DateFilter().get_filter_terms('multi word head dt>"today" dt:"1984-01-01"')
+    assert dtrange_match == ["dt>'today'", "dt:'1984-01-01'"]
+
+    dtrange_match = DateFilter().get_filter_terms('head dt>"today" dt:"1984-01-01" multi word tail')
+    assert dtrange_match == ["dt>'today'", "dt:'1984-01-01'"]
+
+    dtrange_match = DateFilter().get_filter_terms('multi word head dt>="today" dt="1984-01-01"')
+    assert dtrange_match == ["dt>='today'", "dt='1984-01-01'"]
+
+    dtrange_match = DateFilter().get_filter_terms('dt<"multi word date" multi word tail')
+    assert dtrange_match == ["dt<'multi word date'"]
+
+    dtrange_match = DateFilter().get_filter_terms('head dt<="multi word date"')
+    assert dtrange_match == ["dt<='multi word date'"]
+
+    dtrange_match = DateFilter().get_filter_terms("head tail")
+    assert dtrange_match == []
+
+
+def test_date_extraction():
+    extracted_dates = DateFilter().extract_dates("")
+    assert extracted_dates == [], "Expected to handle empty string"
+
+    extracted_dates = DateFilter().extract_dates("head tail")
+    assert extracted_dates == [], "Expected to handle no dates"
+
+    extracted_dates = DateFilter().extract_dates("head CREATED: today tail")
+    assert extracted_dates == [], "Expected relative date to be ignored"
+
+    extracted_dates = DateFilter().extract_dates("head CREATED: today SCHEDULED: 1984-04-01 tail")
+    assert extracted_dates == [datetime(1984, 4, 1, 0, 0, 0)], "Expected only Y-m-d structured date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head SCHEDULED: 01-04-1984 tail")
+    assert extracted_dates == [datetime(1984, 4, 1, 0, 0, 0)], "Expected d-m-Y structured date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head CREATED: today SCHEDULED: 1984/04/01 tail")
+    assert extracted_dates == [datetime(1984, 4, 1, 0, 0, 0)], "Expected only Y/m/d structured date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head SCHEDULED: 01/04/1984 tail")
+    assert extracted_dates == [datetime(1984, 4, 1, 0, 0, 0)], "Expected d/m/Y structured date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head DEADLINE: 01.04.1984 tail")
+    assert extracted_dates == [datetime(1984, 4, 1, 0, 0, 0)], "Expected d.m.Y structured date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("CLOCK: [1984-04-01 Sun 09:50]--[1984-04-01 Sun 10:10] => 24:20")
+    assert extracted_dates == [
+        datetime(1984, 4, 1, 0, 0, 0)
+    ], "Expected single deduplicated date extracted from logbook entry"
+
+    extracted_dates = DateFilter().extract_dates("CLOCK: [1984/03/31 mer 09:50]--[1984/04/01 mer 10:10] => 24:20")
+    expected_dates = [datetime(1984, 4, 1, 0, 0, 0), datetime(1984, 3, 31, 0, 0, 0)]
+    assert all(
+        [dt in extracted_dates for dt in expected_dates]
+    ), "Expected multiple different dates extracted from logbook entry"
+
+
+def test_natual_date_extraction():
+    extracted_dates = DateFilter().extract_dates("head 1 April 1984 tail")
+    assert datetime(1984, 4, 1, 0, 0, 0) in extracted_dates, "Expected natural date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head 1st April 1984 tail")
+    assert datetime(1984, 4, 1, 0, 0, 0) in extracted_dates, "Expected natural date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head 2nd Apr 1984 tail")
+    assert datetime(1984, 4, 2, 0, 0, 0) in extracted_dates, "Expected natural date with short month to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head 4th Apr 1984 tail")
+    assert datetime(1984, 4, 4, 0, 0, 0) in extracted_dates, "Expected natural date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head 11th april 1984 tail")
+    assert (
+        datetime(1984, 4, 11, 0, 0, 0) in extracted_dates
+    ), "Expected natural date with lowercase month to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head 23rd april 84 tail")
+    assert datetime(1984, 4, 23, 0, 0, 0) in extracted_dates, "Expected natural date with 2-digit year to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head 31st march 84 tail")
+    assert datetime(1984, 3, 31, 0, 0, 0) in extracted_dates, "Expected natural date with 2-digit year to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head April 1984 tail")
+    assert extracted_dates == [datetime(1984, 4, 1, 0, 0, 0)], "Expected partial natural date to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head Apr 1984 tail")
+    assert extracted_dates == [
+        datetime(1984, 4, 1, 0, 0, 0)
+    ], "Expected partial natural date with short month to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head apr 1984 tail")
+    assert extracted_dates == [
+        datetime(1984, 4, 1, 0, 0, 0)
+    ], "Expected partial natural date with lowercase month to be extracted"
+
+    extracted_dates = DateFilter().extract_dates("head apr 84 tail")
+    assert extracted_dates == [
+        datetime(1984, 4, 1, 0, 0, 0)
+    ], "Expected partial natural date with 2-digit year to be extracted"
